@@ -1,6 +1,6 @@
 {
   "name": "create-t3-app",
-  "version": "7.5.6",
+  "version": "7.10.1",
   "description": "Create web application with the t3 stack",
   "license": "MIT",
   "repository": {
@@ -43,34 +43,35 @@
   "dependencies": {
     "chalk": "5.2.0",
     "commander": "^10.0.0",
-    "execa": "^7.0.0",
-    "fs-extra": "^11.1.0",
+    "execa": "^7.1.1",
+    "fs-extra": "^11.1.1",
     "gradient-string": "^2.0.2",
-    "inquirer": "^9.1.4",
+    "inquirer": "^9.1.5",
     "ora": "6.1.2",
     "sort-package-json": "^2.4.1"
   },
   "devDependencies": {
-    "@prisma/client": "^4.10.1",
-    "@tanstack/react-query": "^4.24.9",
-    "@trpc/client": "^10.12.0",
-    "@trpc/next": "^10.12.0",
-    "@trpc/react-query": "^10.12.0",
-    "@trpc/server": "^10.12.0",
+    "@prisma/client": "^4.11.0",
+    "@tanstack/react-query": "^4.28.0",
+    "@trpc/client": "^10.18.0",
+    "@trpc/next": "^10.18.0",
+    "@trpc/react-query": "^10.18.0",
+    "@trpc/server": "^10.18.0",
     "@types/fs-extra": "^11.0.1",
     "@types/gradient-string": "^1.1.2",
     "@types/inquirer": "^9.0.3",
-    "@types/node": "^18.14.0",
-    "next": "^13.1.6",
-    "next-auth": "^4.19.2",
-    "prettier": "^2.8.4",
-    "prettier-plugin-tailwindcss": "^0.2.3",
-    "prisma": "^4.10.1",
+    "@types/node": "^18.15.5",
+    "next": "^13.2.4",
+    "next-auth": "^4.20.1",
+    "prettier": "^2.8.6",
+    "prettier-plugin-tailwindcss": "^0.2.6",
+    "prisma": "^4.11.0",
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "superjson": "^1.12.2",
-    "tsup": "^6.6.3",
-    "type-fest": "^3.6.0",
-    "typescript": "^4.9.5"
+    "tsup": "^6.7.0",
+    "type-fest": "^3.7.0",
+    "typescript": "^5.0.2",
+    "zod": "^3.21.4"
   }
 }
