@@ -24,7 +24,7 @@
   "engines": {
     "node": ">=14.16"
   },
-  "packageManager": "pnpm@7.13.0",
+  "packageManager": "pnpm@8.1.0",
   "scripts": {
     "typecheck": "turbo typecheck",
     "build:cli": "turbo --filter=create-t3-app build",
@@ -46,22 +46,24 @@
   },
   "dependencies": {
     "@changesets/changelog-github": "^0.4.8",
-    "@changesets/cli": "^2.26.0",
+    "@changesets/cli": "^2.26.1",
     "@manypkg/cli": "^0.20.0",
-    "@types/eslint": "^8.21.1",
-    "@types/node": "^18.14.0",
-    "@typescript-eslint/eslint-plugin": "^5.53.0",
-    "@typescript-eslint/parser": "^5.53.0",
-    "eslint": "^8.34.0",
-    "eslint-config-prettier": "^8.6.0",
-    "eslint-config-turbo": "^0.0.8",
+    "@total-typescript/ts-reset": "^0.3.7",
+    "@types/eslint": "^8.21.3",
+    "@types/node": "^18.15.5",
+    "@types/prettier": "^2.7.2",
+    "@typescript-eslint/eslint-plugin": "^5.56.0",
+    "@typescript-eslint/parser": "^5.56.0",
+    "eslint": "^8.36.0",
+    "eslint-config-prettier": "^8.8.0",
+    "eslint-config-turbo": "^0.0.9",
     "eslint-plugin-import": "^2.27.5",
-    "eslint-plugin-isaacscript": "^2.4.0",
+    "eslint-plugin-isaacscript": "^2.6.7",
     "eslint-plugin-prettier": "^4.2.1",
     "eslint-plugin-react": "^7.32.2",
     "eslint-plugin-react-hooks": "^4.6.0",
-    "prettier": "^2.8.4",
-    "turbo": "^1.8.1",
-    "typescript": "^4.9.5"
+    "prettier": "^2.8.6",
+    "turbo": "^1.8.5",
+    "typescript": "^5.0.2"
   }
 }
