@@ -8,8 +8,8 @@
     "start": "astro dev",
     "check": "astro check",
     "clean": "rm -rf node_modules .turbo dist .eslintcache",
-    "format": "prettier --write --plugin-search-dir=. **/*.{cjs,mjs,ts,tsx,md,json,astro} --ignore-path ../.gitignore --ignore-unknown --no-error-on-unmatched-pattern",
-    "format:check": "pnpm format --check",
+    "format": "pnpm format:check --write",
+    "format:check": "prettier --check --plugin-search-dir=. **/*.{cjs,mjs,ts,tsx,md,json,astro} --ignore-path ../.gitignore --ignore-unknown --no-error-on-unmatched-pattern",
     "lint": "eslint . --report-unused-disable-directives",
     "lint:fix": "pnpm lint --fix",
     "build": "astro build",
@@ -17,39 +17,39 @@
     "astro": "astro"
   },
   "dependencies": {
-    "@algolia/client-search": "^4.14.3",
-    "@astrojs/image": "^0.15.0",
-    "@astrojs/mdx": "^0.17.0",
-    "@astrojs/sitemap": "^1.0.1",
+    "@algolia/client-search": "^4.15.0",
+    "@astrojs/image": "^0.15.1",
+    "@astrojs/mdx": "^0.17.2",
+    "@astrojs/sitemap": "^1.2.1",
     "@docsearch/css": "^3.3.3",
     "@docsearch/react": "^3.3.3",
     "@fontsource/inter": "^4.5.15",
     "@fontsource/jetbrains-mono": "^4.5.12",
-    "@headlessui/react": "^1.7.11",
-    "@vercel/analytics": "^0.1.10",
+    "@headlessui/react": "^1.7.13",
+    "@vercel/analytics": "^0.1.11",
     "clsx": "^1.2.1",
-    "embla-carousel": "^7.0.9",
-    "embla-carousel-autoplay": "^7.0.9",
+    "embla-carousel": "^7.1.0",
+    "embla-carousel-autoplay": "^7.1.0",
     "sharp": "^0.31.3",
     "tailwind-scrollbar": "^2.1.0",
     "treeify": "^1.1.0"
   },
   "devDependencies": {
-    "@astrojs/react": "^2.0.2",
-    "@astrojs/tailwind": "^3.0.1",
-    "@types/node": "^18.14.0",
+    "@astrojs/react": "^2.1.0",
+    "@astrojs/tailwind": "^3.1.1",
+    "@types/node": "^18.15.5",
     "@types/react": "^18.0.28",
     "@types/react-dom": "^18.0.11",
     "@types/react-typist": "^2.0.3",
     "@types/treeify": "^1.0.0",
-    "@typescript-eslint/parser": "^5.53.0",
-    "astro": "^2.0.14",
-    "eslint": "^8.34.0",
+    "@typescript-eslint/parser": "^5.56.0",
+    "astro": "^2.1.5",
+    "eslint": "^8.36.0",
     "eslint-plugin-astro": "^0.23.0",
     "eslint-plugin-jsx-a11y": "^6.7.1",
-    "prettier": "^2.8.4",
+    "prettier": "^2.8.6",
     "prettier-plugin-astro": "^0.8.0",
-    "prettier-plugin-tailwindcss": "^0.2.3",
+    "prettier-plugin-tailwindcss": "^0.2.6",
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "react-typist": "^2.0.5",
@@ -57,7 +57,7 @@
     "rehype-external-links": "^2.0.1",
     "rehype-slug": "^5.1.0",
     "remark-code-titles": "^0.1.2",
-    "tailwindcss": "^3.2.7",
-    "typescript": "^4.9.5"
+    "tailwindcss": "^3.3.0",
+    "typescript": "^5.0.2"
   }
 }
