@@ -1,6 +1,8 @@
 /** @type {import('prettier').Config} */
-module.exports = {
+const config = {
   ...require("../.prettierrc.cjs"),
   plugins: [require.resolve("prettier-plugin-tailwindcss")],
-  tailwindConfig: "./template/extras/config/tailwind.config.cjs",
+  tailwindConfig: "./template/extras/config/tailwind.config.ts",
 };
+
+module.exports = config;
