@@ -67,15 +67,15 @@ const main = async () => {
     await installDependencies({ projectDir });
   }
 
-  if (!noGit) {
-    await initializeGit(projectDir);
-  }
-
   // Rename _eslintrc.json to .eslintrc.json - we use _eslintrc.json to avoid conflicts with the monorepos linter
   fs.renameSync(
     path.join(projectDir, "_eslintrc.cjs"),
     path.join(projectDir, ".eslintrc.cjs"),
   );
+
+  if (!noGit) {
+    await initializeGit(projectDir);
+  }
 
   logNextSteps({ projectName: appDir, packages: usePackages, noInstall });
 
