@@ -1,5 +1,5 @@
 /** @type {import('prettier').Config} */
-module.exports = {
+const config = {
   ...require("../.prettierrc.cjs"),
   plugins: [
     require.resolve("prettier-plugin-astro"),
@@ -15,5 +15,7 @@ module.exports = {
     },
   ],
   astroAllowShorthand: false,
-  tailwindConfig: "./tailwind.config.cjs",
+  tailwindConfig: "./tailwind.config.ts",
 };
+
+module.exports = config;
