@@ -1,13 +1,12 @@
-import { defineConfig } from "astro/config";
-import react from "@astrojs/react";
-import tailwind from "@astrojs/tailwind";
 import image from "@astrojs/image";
 import mdx from "@astrojs/mdx";
-import rehypeExternalLinks from "rehype-external-links";
-import remarkCodeTitles from "remark-code-titles";
-import rehypeSlug from "rehype-slug";
-import rehypeAutolinkHeadings from "rehype-autolink-headings";
+import react from "@astrojs/react";
 import sitemap from "@astrojs/sitemap";
+import { defineConfig } from "astro/config";
+import rehypeAutolinkHeadings from "rehype-autolink-headings";
+import rehypeExternalLinks from "rehype-external-links";
+import rehypeSlug from "rehype-slug";
+import remarkCodeTitles from "remark-code-titles";
 
 /** @link https://astro.build/config */
 export default defineConfig({
@@ -45,7 +44,6 @@ export default defineConfig({
   },
   integrations: [
     react(),
-    tailwind({ config: { applyBaseStyles: false } }),
     image({
       serviceEntryPoint: "@astrojs/image/sharp",
     }),
